# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Prahari API
  version: 2.0.0
  summary: Alert correlation, cohesion, business-risk ranking, counterfactuals, breach clocks and adversarial evaluation.
  license: { name: MIT, identifier: MIT }
  description: |
    Contract-first API for Prahari (Microsoft Innovate 2026, problem #25, absorbing #22 and #27).

    * All timestamps are UTC, RFC 3339. The UI renders IST.
    * All errors are RFC 9457 `application/problem+json` — switch on `code`, never on `title`.
    * List endpoints use cursor pagination and return `{ data, page }`.
    * Live progress is Server-Sent Events. `EventSource` cannot set headers, so SSE
      endpoints also accept `?access_token=`; only those endpoints do.
    * Incidents are immutable per run and identified by `(run_id, incident_id)`.
      Analyst state is keyed on the stable `incident_id` alone, so verdicts survive a re-run.
    * Clients must ignore unknown fields: additive changes stay in v1.
    * Human-readable companion: `api-contract.md`. Design: `design.md`.
servers:
  - { url: 'http://localhost:8080/api/v1', description: Docker Compose (via nginx) }
  - { url: 'http://localhost:8000/api/v1', description: Go API directly }
security: [ { bearerAuth: [] } ]
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Exchange username and password for a JWT
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Current user
      responses:
        '200': { description: The authenticated user, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────────── System
  /meta:
    get:
      tags: [System]
      operationId: getMeta
      summary: Build, storage dialect, ATT&CK version and active engine configuration
      responses:
        '200': { description: Metadata, content: { application/json: { schema: { $ref: '#/components/schemas/Meta' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────────── Datasets
  /simulations:
    post:
      tags: [Datasets]
      operationId: createSimulation
      summary: Generate a synthetic dataset (auth events, alerts, hidden ground truth) and ingest it
      description: Requires role `lead`. Deterministic for a given seed and parameters.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SimulationRequest' } } }
      responses:
        '201':
          description: Dataset created and ingested
          headers: { Location: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Dataset' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /datasets:
    get:
      tags: [Datasets]
      operationId: listDatasets
      summary: List datasets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: kind
          in: query
          schema: { type: string, enum: [simulated, ingested, adversarial] }
      responses:
        '200':
          description: Datasets, newest first
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Dataset' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Datasets]
      operationId: createDataset
      summary: Register an empty dataset for external NDJSON ingest (role `lead`)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DatasetCreate' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Dataset' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /datasets/{datasetId}:
    parameters: [ { $ref: '#/components/parameters/DatasetIdPath' } ]
    get:
      tags: [Datasets]
      operationId: getDataset
      summary: Get a dataset
      responses:
        '200': { description: Dataset, content: { application/json: { schema: { $ref: '#/components/schemas/Dataset' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /datasets/{datasetId}/auth-events:
    parameters: [ { $ref: '#/components/parameters/DatasetIdPath' } ]
    post:
      tags: [Datasets]
      operationId: ingestAuthEvents
      summary: Ingest raw authentication events (NDJSON); detectors raise alerts from them
      description: |
        Requires role `lead`. One `AuthEventInput` per line. Max body 50 MB, max line 1 MB.
        Invalid lines are skipped and reported; duplicate `event_id` values are ignored.
      requestBody:
        required: true
        content:
          application/x-ndjson:
            schema: { $ref: '#/components/schemas/AuthEventInput' }
            example: |
              {"event_id":"AE-1","timestamp":"2026-09-29T03:32:11Z","username":"priya@corp.local","src_ip":"185.220.101.7","geo":"NL-AMS","result":"failure","app":"vpn"}
      responses:
        '200': { description: Ingest report, content: { application/json: { schema: { $ref: '#/components/schemas/IngestReport' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /datasets/{datasetId}/alerts:
    parameters: [ { $ref: '#/components/parameters/DatasetIdPath' } ]
    post:
      tags: [Datasets]
      operationId: ingestAlerts
      summary: Ingest alerts from any source (NDJSON, role `lead`)
      requestBody:
        required: true
        content:
          application/x-ndjson:
            schema: { $ref: '#/components/schemas/AlertInput' }
      responses:
        '200': { description: Ingest report, content: { application/json: { schema: { $ref: '#/components/schemas/IngestReport' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }

  # ───────────────────────────────── Alerts
  /alerts:
    get:
      tags: [Alerts]
      operationId: listAlerts
      summary: Raw alert feed (the firehose)
      parameters:
        - $ref: '#/components/parameters/DatasetIdQuery'
        - { name: from,     in: query, schema: { type: string, format: date-time } }
        - { name: to,       in: query, schema: { type: string, format: date-time } }
        - { name: source,   in: query, description: Comma-separated, schema: { type: string, example: 'auth,edr' } }
        - { name: severity, in: query, description: Comma-separated, schema: { type: string, example: 'high,critical' } }
        - { name: rule_id,  in: query, schema: { type: string } }
        - { name: entity,   in: query, description: 'Entity key, e.g. host:dc01', schema: { type: string } }
        - { name: q,        in: query, description: Case-insensitive substring on rule name, schema: { type: string } }
        - { name: sort,     in: query, schema: { type: string, enum: [ts, -ts], default: ts } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Alert' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /alerts/stats:
    get:
      tags: [Alerts]
      operationId: getAlertStats
      summary: Counts by source, severity and hour
      parameters: [ { $ref: '#/components/parameters/DatasetIdQuery' } ]
      responses:
        '200': { description: Stats, content: { application/json: { schema: { $ref: '#/components/schemas/AlertStats' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /alerts/{alertId}:
    get:
      tags: [Alerts]
      operationId: getAlert
      summary: One alert including its raw payload
      parameters:
        - { name: alertId, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/DatasetIdQuery'
      responses:
        '200': { description: Alert with raw payload, content: { application/json: { schema: { $ref: '#/components/schemas/AlertWithRaw' } } } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────────────── Runs
  /runs:
    get:
      tags: [Runs]
      operationId: listRuns
      summary: Run history
      parameters:
        - { name: dataset_id, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Runs, newest first
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Run' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Runs]
      operationId: createRun
      summary: Start a correlation run (asynchronous, role `lead`)
      description: |
        Only one run per dataset may be active. A second request returns `409 RUN_IN_PROGRESS`
        carrying `active_run_id`; the client should attach to that run's SSE stream.
        Exclusivity uses a `run_locks` row with a TTL, which behaves identically on both
        storage dialects and survives a process crash.
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RunCreate' } } }
      responses:
        '202':
          description: Run accepted
          headers: { Location: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /runs/{runId}:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    get:
      tags: [Runs]
      operationId: getRun
      summary: Run status and summary
      responses:
        '200': { description: Run, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /runs/{runId}/cancel:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    post:
      tags: [Runs]
      operationId: cancelRun
      summary: Cancel a queued or running run (role `lead`)
      responses:
        '202': { description: Cancellation requested, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /runs/{runId}/receipt:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    get:
      tags: [Runs]
      operationId: getReceipt
      summary: Determinism receipt for a run
      description: |
        Two runs with the same `input_hash` and `config_hash` produce the same `output_hash`
        — on either storage dialect, on any machine. This proves the engine is a pure function
        of its inputs. It does NOT prove the answer is correct; that is what `/evaluations`
        and `/adversary` are for.

        `config` includes fingerprints of the CMDB, rule statistics and active suppressions
        as the run saw them, because those also change the output.
      responses:
        '200': { description: Receipt, content: { application/json: { schema: { $ref: '#/components/schemas/Receipt' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /runs/{runId}/events:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    get:
      tags: [Runs, Events]
      operationId: streamRunEvents
      summary: Live run progress (Server-Sent Events)
      description: |
        Event types: `status` (RunStatusEvent), `stage` (RunStageEvent), `done` (RunDoneEvent),
        `error` (RunErrorEvent). Heartbeat comment `: ping` every 15 s. Resume with
        `Last-Event-ID`; the last 200 events per run are replayed. Connecting to a finished
        run replays then closes.
      parameters:
        - $ref: '#/components/parameters/LastEventId'
        - $ref: '#/components/parameters/AccessTokenQuery'
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/RunStatusEvent' }
                  - { $ref: '#/components/schemas/RunStageEvent' }
                  - { $ref: '#/components/schemas/RunDoneEvent' }
                  - { $ref: '#/components/schemas/RunErrorEvent' }
              example: |
                id: 41
                event: stage
                data: {"run_id":"0192a7c4-6e1b-7c3e-9a51-2f4d7a0c1b11","stage":"group","count":28,"elapsed_ms":61}
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /events:
    get:
      tags: [Events]
      operationId: streamGlobalEvents
      summary: Global notifications (Server-Sent Events)
      parameters:
        - $ref: '#/components/parameters/LastEventId'
        - $ref: '#/components/parameters/AccessTokenQuery'
      responses:
        '200':
          description: Event stream
          content: { text/event-stream: { schema: { $ref: '#/components/schemas/GlobalEvent' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────────── Incidents
  /incidents:
    get:
      tags: [Incidents]
      operationId: listIncidents
      summary: Ranked incident queue
      description: '`-risk` (default) is rank order: priority band first, then risk, so a P1 floor sits above higher-scoring P2s.'
      parameters:
        - $ref: '#/components/parameters/DatasetIdQuery'
        - { name: run_id,   in: query, description: Defaults to the dataset's current run, schema: { type: string, format: uuid } }
        - { name: priority, in: query, description: 'Comma-separated, e.g. P1,P2', schema: { type: string } }
        - { name: label,    in: query, schema: { type: string } }
        - { name: status,   in: query, schema: { type: string } }
        - { name: cohesion, in: query, description: Comma-separated, schema: { type: string, example: fragile } }
        - { name: asset,    in: query, schema: { type: string } }
        - { name: assignee, in: query, schema: { type: string } }
        - { name: breach,   in: query, schema: { type: boolean } }
        - { name: sort,     in: query, schema: { type: string, enum: ['-risk', risk, '-first_seen', first_seen], default: '-risk' } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Incidents
          headers: { ETag: { schema: { type: string } } }
          content:
            application/json:
              schema:
                type: object
                required: [data, page, run_id, totals, thresholds]
                properties:
                  run_id:     { type: string, format: uuid }
                  totals:     { $ref: '#/components/schemas/IncidentTotals' }
                  thresholds: { $ref: '#/components/schemas/Bands' }
                  data:       { type: array, items: { $ref: '#/components/schemas/IncidentSummary' } }
                  page:       { $ref: '#/components/schemas/PageInfo' }
        '304': { description: Not modified }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: getIncident
      summary: Incident detail with risk breakdown and cohesion
      responses:
        '200':
          description: Incident
          headers: { ETag: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/IncidentDetail' } } }
        '304': { description: Not modified }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Incidents]
      operationId: updateIncident
      summary: Change workflow status or assignee
      description: Writes to `incident_state`, which is keyed on the stable incident id, so the change survives a re-run.
      parameters:
        - { name: If-Match, in: header, required: true, description: ETag from the last GET, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IncidentPatch' } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/IncidentDetail' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /incidents/{incidentId}/alerts:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: listIncidentAlerts
      summary: Member alerts in time order
      parameters:
        - { name: on_chain, in: query, description: Only alerts on the forward kill-chain path, schema: { type: boolean } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/IncidentAlert' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}/graph:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: getIncidentGraph
      summary: Entity graph projection (for React Flow)
      description: Edges between the same two entities are collapsed to one (the on-chain or bridge alert wins); at most 300 edges.
      responses:
        '200': { description: Graph, content: { application/json: { schema: { $ref: '#/components/schemas/IncidentGraph' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}/timeline:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: getIncidentTimeline
      summary: Kill-chain swim-lane projection
      responses:
        '200': { description: Timeline, content: { application/json: { schema: { $ref: '#/components/schemas/IncidentTimeline' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}/narrative:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: getNarrative
      summary: Grounded incident brief (cached, LLM, or deterministic template)
      description: |
        Never fails because of the model. On timeout, error, open circuit breaker or failed
        citation validation, a deterministic template narrative is returned with `source: template`.
      responses:
        '200': { description: Narrative, content: { application/json: { schema: { $ref: '#/components/schemas/Narrative' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}/narrative/regenerate:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    post:
      tags: [Incidents]
      operationId: regenerateNarrative
      summary: Bypass the cache and generate again (10/min per user)
      responses:
        '200': { description: Narrative, content: { application/json: { schema: { $ref: '#/components/schemas/Narrative' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /incidents/{incidentId}/cohesion:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: getCohesion
      summary: How confident the system is that these alerts are ONE incident
      description: |
        Bridges are found with Tarjan's low-link DFS in O(V+E) over alert-to-alert links (two
        alerts sharing several entities are one link). An incident is `fragile` when a bridge
        rests on a low-weight (common) entity, removing it produces two substantial halves,
        and the halves look different (entity-set Jaccard below 0.5; halves that look alike
        are one recurring activity split by time). `split_preview` is present only when `fragile`.
      responses:
        '200': { description: Cohesion, content: { application/json: { schema: { $ref: '#/components/schemas/Cohesion' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}/split:
    parameters: [ { $ref: '#/components/parameters/IncidentIdPath' } ]
    post:
      tags: [Incidents]
      operationId: splitIncident
      summary: Materialise a fragile incident as two incidents (role `lead`)
      description: |
        Creates a NEW run rather than editing the existing one, so the original stays immutable
        and its determinism receipt stays valid. The cut is recorded on the dataset (and in the
        new run's config hash) so later runs honour it. The analyst's judgement is recorded in
        the audit log.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SplitRequest' } } }
      responses:
        '201': { description: Split applied, content: { application/json: { schema: { $ref: '#/components/schemas/SplitResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /incidents/{incidentId}/counterfactuals:
    parameters:
      - $ref: '#/components/parameters/IncidentIdPath'
      - $ref: '#/components/parameters/RunIdQuery'
    get:
      tags: [Incidents]
      operationId: getCounterfactuals
      summary: The minimal changes that would flip this incident's priority
      description: |
        Computed on demand from the closed-form score; nothing is stored. `what_if_criticality`
        previews a CMDB edit without persisting it — this is what the Risk tab's live slider calls.
        A `boundary` entry with `reachable: false` means that factor alone cannot reach the band
        edge even at its extreme (or a hard P1 floor holds); it is reported honestly rather than clamped.
      parameters:
        - name: what_if_criticality
          in: query
          description: 'Preview a criticality change, e.g. `dc01:4`. Repeatable.'
          schema: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}:(10|[1-9])$' }
      responses:
        '200': { description: Counterfactuals, content: { application/json: { schema: { $ref: '#/components/schemas/CounterfactualSet' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /adversary/campaigns/{campaignId}:
    parameters: [ { $ref: '#/components/parameters/CampaignIdPath' } ]
    get:
      tags: [Adversary]
      operationId: getCampaign
      summary: Campaign with per-budget results
      responses:
        '200': { description: Campaign, content: { application/json: { schema: { $ref: '#/components/schemas/CampaignDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /adversary/campaigns/{campaignId}/events:
    parameters: [ { $ref: '#/components/parameters/CampaignIdPath' } ]
    get:
      tags: [Adversary, Events]
      operationId: streamCampaignEvents
      summary: Live campaign progress (Server-Sent Events)
      description: 'Event types: `budget.started`, `budget.finished`, `done`, `error`.'
      parameters:
        - $ref: '#/components/parameters/LastEventId'
        - $ref: '#/components/parameters/AccessTokenQuery'
      responses:
        '200':
          description: Event stream
          content: { text/event-stream: { schema: { $ref: '#/components/schemas/CampaignEvent' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /rules:
    get:
      tags: [Rules]
      operationId: listRules
      summary: Detection rules with learned confidence
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { type: array, items: { $ref: '#/components/schemas/RuleStat' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    BadRequest:           { description: 'VALIDATION_FAILED, MALFORMED_JSON, INVALID_CURSOR', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 128 }
...